pyproject.toml
[project]
name = "pykuma"
version = "0.1.0"
description = "PyKuma - A frame-accurate Street Fighter III: 3rd Strike recreation in Python with Pydantic validation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygame>=2.6.1",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "typing-extensions>=4.8.0",
]
authors = [
    { name = "ShirkAttack" }
]
license = { text = "MIT" }
keywords = ["fighting-game", "street-fighter", "pygame", "game-development", "pydantic", "sf3", "3rd-strike"]

[project.scripts]
sf3 = "street_fighter_3rd.main:main"
sf3-menu = "street_fighter_3rd.main_with_menu:main"
sf3-training = "street_fighter_3rd.main_with_menu:main_training"
sf3-dev = "street_fighter_3rd.main_with_menu:main_dev"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/street_fighter_3rd"]
